add day 19 solutions
Wow, big jump in difficulty. I snagged 12th on the leaderboard, but
only because I got lucky: my naive algorithm produced the correct
answer. It's a greedy algorithm that reduces the string as much as
possible, paying no attention to what order the reductions are
applied in. But apparently that's good enough!