add day 20 solutions
had a hard time with this one :(
I didn't think to trade space for time, so my initial attempts all
ran too slowly. I also made the mistake of using Haskell at first
because it seemed better-suited to the problem, but without O(1)
list access it was dead in the water.