add day 19 solutions
I really struggled to find an efficient algorithm for part 2.
Eventually I gave up and generated the sequence using a slow (but
correct) approach. I was surprised at how regular it was. It's
deceptively random-looking for low n.

I'm not sure why it produces this sequence, but I'm glad that I was
able to find a constant-time algorithm.