add day 20 solutions
Another poor placement. I made the mistake of thinking that a naive
solution would be too slow for part 2. In fact, it does just fine
with some simple optimizations (namely, sorting the ranges and
skipping to the end of the ban). Instead, I went down a rabbit hole
of trying to merge overlapping ranges together, which I never got
properly working. The idea was that once all the ranges were merged,
you could simply iterate over the bans and subtract their ranges
from the total. One of the "gotchas" in this problem was avoiding
uint32 wrap-around; it definitely burned me at least once.