add day 23 solutions
I finished the first part very quickly, but the second part gave
me a strange amount of trouble. I think I was just tired.
I like the approach of actually adding a mul instruction, but
note that its "toggle" behavior is undefined. I also considered
hard-coding the multiplication to happen at a specific instruction
offset, but that felt just a bit too hacky.

Note that it would also be trivial to determine the answer via
pattern matching. By substituting different values for a, it
becomes apparent that the output is proportional to factorial(a).
After the factorial is calculated, a final product is added to the
total; in my case 89*77. This method was likely used by some of
the fastest finishers.