add day 24 solutions
Too slow. I should really add a helper function for finding the
shortest path in a grid. I got burned because my algorithm only
checked that it had visited a node before, not that the current
distance was greater as well. So my initial answers were all way
too high.