add day 3 solutions
I really didn't want to write a program to fill in the spiral,
because it seemed like a pain to figure out which index came
next. Instead, I noticed that the bottom-right diagonal of the
spiral contained the squares of odd numbers (1, 9, 25, 49...).
Exploiting this, I located the odd square closest to my input,
which turned out to be 527. The x and y coordinates of this
square are (n-1)/2 = 263. Then, I subtracted my input from 527^2,
yielding 51. Since the length of the edge was 263, and 51*2 < 263,
the final distance was simply 263+263-51.

I kinda cheated on the second part. Again, I didn't want to fill
in the spiral, so instead I just plugged the sequence into OEIS,
which helpfully provided a listing of all the sequence values up
to 961. All I had to do was locate the first value in the sequence
larger than my input.