add day 15 solutions
I didn't get to start this one at midnight. Shame, because I'm
sure I could have finished it quickly.
There's a more elegant solution that uses channels to filter
the generator values, but it runs too slowly to be practical.
This would be a fun one to rewrite in ply.