add day 15 solutions
I really hated this problem. I had a solution that worked on all
of my test cases within 90 mins or so, which would have landed me
comfortably in the top 100. But there was a bug that only affected
my input. I finally fixed it two hours later. The next day I decided
to ruthlessly hate-optimize it, and as a result it now runs over a
hundred times faster.