add day 19 solutions
man, I knew there would be a problem like this -- manually "jetting"
assembly code. I guess it's more general than that: there are quite
a few Advent problems where part 1 can be simulated, but part 2
requires you to find a pattern and "cheat" to figure out where it
will be on a particular step. So the meta-solution is: write the
simulation code, use it for part 1, then for part 2, print the output
on each iteration, look for a pattern, then exploit the pattern.
Importantly, you don't need to be "perfect" here; you just need to
exploit the pattern to a degree that your code runs fast enough.
Like for this problem, I didn't need to actually understand the full
computation (I imagine it was something like "sum all the factors of
this large number"), I just needed to speed up the hottest loop.