add day 2 solutions
I'm surprised to see a "virtual machine" problem so early in the
month -- usually those don't come until later. At least I didn't
have to read any assembly code this time. I placed well on part 1,
but lost a lot of time on part 2 for stupid reasons. It's so hard
to walk the line between speed and attention to detail!